+++ app/src/main/java/com/sessac/myaitrip/fragment/SignupProfileFragment.kt
@@ -5,6 +5,7 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import com.sessac.myaitrip.databinding.FragmentSignupProfileBinding
+import com.sessac.myaitrip.presentation.common.ViewBindingBaseFragment
 
 /**
  * 프로필 추가/수정 페이지
