+++ app/src/main/java/com/sessac/myaitrip/fragment/ViewBindingBaseFragment.kt
@@ -1,31 +1,7 @@
 package com.sessac.myaitrip.fragment
 
-import android.os.Bundle
 import android.view.LayoutInflater
-import android.view.View
 import android.view.ViewGroup
-import androidx.fragment.app.Fragment
-import androidx.viewbinding.ViewBinding
 
 typealias FragmentInflate<T> = (LayoutInflater, ViewGroup?, Boolean) -> T
 
-abstract class ViewBindingBaseFragment<VB: ViewBinding>
-    (private val inflate: FragmentInflate<VB>): Fragment() {
-
-    private var _binding: VB? = null
-    val binding get() = _binding!!
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        _binding = inflate.invoke(inflater, container, false)
-        return binding.root
-    }
-
-    override fun onDestroyView() {
-        super.onDestroyView()
-        _binding = null
-    }
-}
